Update README.md
This project demonstrates the use of deep learning, specifically Convolutional Neural Networks (CNNs), to recognize handwritten digits using the famous MNIST dataset. The MNIST dataset consists of 70,000 grayscale images of handwritten digits (0–9), each of size 28x28 pixels. These images are widely used as a benchmark in the fields of computer vision and deep learning.

The objective of the project is to build an image classification model that accurately identifies digits from the input images. To achieve this, a CNN model is designed using TensorFlow and Keras, which is trained on 60,000 images and evaluated on 10,000 test images. The architecture includes convolutional layers to extract features, pooling layers for downsampling, and dense layers to perform classification. The final output layer uses the softmax activation function to predict probabilities for each of the 10 digit classes.